colors fixed & pixel arrays removed
