+++ Makefile
@@ -6,7 +6,7 @@
 #    By: amakela <amakela@student.hive.fi>          +#+  +:+       +#+         #
 #                                                 +#+#+#+#+#+   +#+            #
 #    Created: 2024/07/24 13:25:09 by amakela           #+#    #+#              #
-#    Updated: 2024/09/26 16:02:55 by amakela          ###   ########.fr        #
+#    Updated: 2024/09/27 16:00:03 by amakela          ###   ########.fr        #
 #                                                                              #
 # **************************************************************************** #
 
@@ -28,8 +28,8 @@ CC			= cc
 
 CFILES		=	src/game.c				src/main.c				src/parse_elements.c	\
 				src/parse_map.c			src/utils_cleaning.c	src/utils_parsing.c		\
-				src/movement.c			src/rotation.c			src/drawing.c			\
-				src/ray_casting.c
+				src/movement.c			src/rotation.c			src/utils_colors.c		\
+				src/ray_casting.c		src/drawing.c
 
 OFILES		= $(CFILES:.c=.o)
